Remove useless trees from Bulwark and Berserker
These are newbie traps.  One of the 2H trees is the old pre-revamp version, the Archery tree doesn't even have a mastery.  These were clearly left in as oversights.

Technically Ogre can make use of some of these trees but it isn't a good idea to leave confusing trees on base classes for gimmick builds on an unlocked race.