Remove Munitions from Archer and readd Poisons
I started revising this tree and after getting pretty far realized it was a futile endeavor.  While this is a cool and thematic tree it has a lot of problems:

- Enormous amount of free power passively

- Problem solves heal sustain, sustain removal, armor/damage penetration, AoE, damage type problems all with the same investment

- Applies a stupid amount of debuffs, up to 3 debuffs passively on hit from just this tree

- Very hard to read what each talent does and Archer is a beginning class

This can't be fixed with number nerfs, it needs to do a lot less stuff than it currently does and that would necessitate a lot of work.  Best to kill it for now.  Poisons is a well designed tree now that fills a similar role.