Improve motion blur and invisibility shader
I have no idea how any of this works, stof just told me what to do and I subjectively evaluated it to be better.  On 1.5 the invisibility shader is very loud making it hard to see things.

stof said that this change was probably a bug fix but I have no idea:
//float factor = ((float(blursize)*2.0)+1.0);
float factor = (blursize+1)*(blursize+1);