Revise Afflicted / Dark Sustenance
- Change Feed to autocast on nearest enemy in combat (huge QOL buff, no real game impact, targeting can still be changed manually)

- Replace Suffuse Life with a life regen drain passive.  This talent is not interesting, and it has a weird breakpoint at TL5 that is pretending to be useful but isn't.  This class is pretty messy as is I opted to just simplify this.

Also, I want to shift more healing to basic regen in places where it fits well so theres more differentiation for things like Curse of Death to play with.

Amusingly enough while editing this I noticed the life steal code was already there unused.  Was this a problem for some reason?