Remove melee retaliation from many NPCs
This stat is sprinkled everywhere and I have no idea why.  A quick search of of my TOME files finds over 200 entries for it.

This stat has some use for differentiating between melee attack styles but in general should be avoided, as its severely punishing and most players don't even notice it exists.  Note that whether they notice it or not it is still getting them killed.

A Normal difficulty Daikara giant can have 15 melee retaliation.  Thats 90 damage on a Flurry.  I do not believe that anyone sat down and decided this was a desirable disadvantage for melee to have, rather whenever something is vaguely elementally themed people just put this stat on them.

This prunes the stat from most things that aren't strongly elementally themed (dragons, elementals).  It probably doesn't go far enough but its a good start.