Fix weapon damage on early NPCs
One consequence of the weapon formula and mastery changes is that the first weapon mastery point is worth far more.  This is good for PCs, but a little over the top on some starting NPCs.

As it turns out this works out well since the weapon NPCs are still stronger than other early things just by virtue of the fact that get to use gear.