Update Worldly Knowledge to only include most of the currently available escort/quest trees
Leaving Light available and such is not a good idea. As this was basically only used to get Light on some builds a token buff of 5 generic points is included, but this prodigy could probably use more love eventually.

Also, this should restructured so that addons and campaigns can easily modify the available talent trees.