Remove Combat Veteran
In the past I've defended this as a tree worth keeping because it lets newbies not worry about stamina management at a cost.  In truth, that is a terrible reason to keep this around.  All 4 talents are both weak and extremely underpowered and they aren't doing anything interesting.  Best to prune this and let Conditioning and such be the "sustained combat" trees.

All effected classes get +1 life/stamina regen base as compensation for the talents they would optimally not even have taken.   This is mostly QOL.