Improve Inner Power and add a periodic damage shield
AB has a bit of the same problem AM did where they're tanky but so fragile on initial engage that they have to play overly safe and tedious.  This buffers that situation a bit.

Beyond that this talent was never worth activating with a mana cost, so now its passive.  This should help open AB actually using more spells.