Nerf Fungus numbers and improve log display
This is basically a giant hack to simulate how delayedLogDamage behaves.  We add up all the turn gain each heal but only push it to the log at the end of ticks *or* on talent use (so it appears instantly on instant use healing talents).  Might still be too spammy.  Might still be overpowered.