Remove the HP bonus on Insane and Nightmare
This can be brought back later, but for the time being high difficulty players are frustrated by the game length and all other solutions are complex (zone overhauls) and untested.  This is guaranteed to speed the game up and with damage being lower in general it should be fine.

Madness is somewhat defined by its massive HP and I've gotten approximately zero feedback on the current state of it, so leaving that alone for now.