Add a combat_precomputed_x property for all of the powers, Accuracy, and Defense that lets you set them to a fixed value
This is mostly to make it easy for pets to inherit a snapshot of their summoner's stats at cast without fancy function capture and such.

Ideally we would just add the summoner's raw power to the summon, but at present there is no way to calculate raw power without information loss.